shamu: introduce a much less crippled liblights.
Each color (RGB) of the tri-color LED sits on it's own GPIO, which
is a rather unique design. As such, the leds-qpnp driver is not
currently developed or designed for robust LED controls. Kernel
dtsi LED configuration will need to be overhauled for a fully
comprehensive solution.

This hal introduces basic user controls including custom charge and
notification colors. Blink and breathe are not supported at this time.

** Modifications for Marshmellow **

include malloc.h

Change-Id: Ibd3063c3ebfda778705a0ab5156ae65489bff68e