Update patch set 1
Patch Set 1:

Script, because that's how it's been in my kernels always, and you can use my kernel flasher to change defaults to your liking, it will regenerate the script (no need to touch ramfs since it's a file withing /system).

As for SELinux, that's how I want it, but sure can be changed to stick to enforcing by default, I just added the script with the settings I use myself.

Patch-set: 1
Label: Code-Review=0