Update patch set 2
Patch Set 2:

Problem is, BOTH of those paths do exist and do work ?!

Also all the following commits need this one in as they mixed the dt2w with the base BoardConfig.mk change which adds them to the ROM.

As long as this isn't fixed, we can't add the rest.

And most of all ... where / how can this be usable ?

Take sun enhancement mode (manualls writing to sysfs works fine) would need a switch (tile ?) in the UI.

This would first need to be added to the ROM-side.

Patch-set: 2