Update patch set 3
Patch Set 3: Code-Review+1

Works now, setting is in "Display Settings"

PS: vibrator strength is present, but gives a Settings FC when used (popup with slider is not shown).

Will need to investigate, but must be a detail.

Sunlight enhancement mode seems to not have an option yet, might need a commit in settings.

Patch-set: 3
Label: Code-Review=+1