Revert "kernel: use in-kernel hotplugging & thermal management"
Keep boot stock (aka. mpdecision and thermal-engine). If Yank555.lu kernel is found, it will be configured as usual but ONLY AFTER boot complete, as it used to be.

Also, no intruding as no configuration will be set if a custom kernel is run (see previous commit).

Both of these ensure proper boot whatever custom kernel is used (if it boots like stock, though !), and custom kernel has to take care to switch to whatever if brings to the table.

This allows for best possible custom kernel support, if using anykernel flash method, else kernel will bring ramfs and all of this doesn't apply anyway.

This reverts commit d46d20522424b74321c59311c163bb96b3a790c7.

Change-Id: I005db5af28c65eb27f638ea951b5ccc51af64c20