shamu: ro.telephony.get_imsi_from_sim=true
Store correct IMSI when retreived from SIMRecords and use it for RuimRecords
This only affects CDMA sims as the prop is read in RuimRecords.
With this the proper mcc mnc is set, thereby getting the proper overlays.
This fixes IMS as well as the proper displaying of the apn settings.

Change-Id: I0a98221e36c1f352f2bf579695be83b73c15cc9f
(cherry picked from commit 47a40a6107b6de45903259edecc02f7cf5038833)